update internal int representation to i8
Were dealing with d6 dice, handling values up to 2 billion seems
mildy overkill

+++ dice_backend/src/ordering/expr.rs
@@ -904,7 +904,7 @@ impl OrdTrait for OrderedExpression {}
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub enum ConstantValue {
     Bool(bool, OrdType),
-    Int(i32, OrdType),
+    Int(i8, OrdType),
 }
 impl AsRef<OrdType> for ConstantValue {
     #[inline(always)]
